frontend Resource: Refactor MainInfoSection into own component
