+++ frontend/src/components/common/Resource/Resource.tsx
@@ -12,7 +12,7 @@ import Editor from '@monaco-editor/react';
 import { Location } from 'history';
 import { Base64 } from 'js-base64';
 import _ from 'lodash';
-import React, { isValidElement, PropsWithChildren } from 'react';
+import React, { PropsWithChildren } from 'react';
 import { useTranslation } from 'react-i18next';
 import { generatePath, NavLinkProps, useLocation } from 'react-router-dom';
 import { labelSelectorToQuery } from '../../../lib/k8s';
@@ -26,27 +26,23 @@ import {
 import Pod, { KubePod } from '../../../lib/k8s/pod';
 import { createRouteURL, RouteURLProps } from '../../../lib/router';
 import { getThemeName } from '../../../lib/themes';
-import { DefaultHeaderAction, HeaderAction } from '../../../redux/actions/actions';
-import { useTypedSelector } from '../../../redux/reducers/reducers';
 import { useHasPreviousRoute } from '../../App/RouteSwitcher';
-import Loader from '../../common/Loader';
 import { SectionBox } from '../../common/SectionBox';
-import SectionHeader, { HeaderStyleProps } from '../../common/SectionHeader';
-import SimpleTable, { NameValueTable, NameValueTableRow } from '../../common/SimpleTable';
+import SectionHeader from '../../common/SectionHeader';
+import SimpleTable, { NameValueTable } from '../../common/SimpleTable';
 import DetailsViewSection from '../../DetailsViewSection';
 import { PodListProps, PodListRenderer } from '../../pod/List';
 import { LightTooltip, ObjectEventList } from '..';
 import Empty from '../EmptyContent';
-import ErrorBoundary from '../ErrorBoundary';
 import { DateLabel, HoverInfoLabel, StatusLabel, StatusLabelProps, ValueLabel } from '../Label';
 import Link, { LinkProps } from '../Link';
 import { useMetadataDisplayStyles } from '.';
-import DeleteButton from './DeleteButton';
-import EditButton from './EditButton';
-import { MetadataDictGrid, MetadataDisplay } from './MetadataDisplay';
+import { MainInfoSection, MainInfoSectionProps } from './MainInfoSection/MainInfoSection';
+import { MetadataDictGrid } from './MetadataDisplay';
 import PortForward from './PortForward';
-import { RestartButton } from './RestartButton';
-import ScaleButton from './ScaleButton';
+
+export { MainInfoSection };
+export type { MainInfoSectionProps };
 
 export interface ResourceLinkProps extends Omit<LinkProps, 'routeName' | 'params'> {
   name?: string;
@@ -67,163 +63,6 @@ export function ResourceLink(props: ResourceLinkProps) {
     <Link routeName={routeName} params={routeParams} state={state}>
       {name}
     </Link>
-  );
-}
-
-export interface MainInfoSectionProps {
-  resource: KubeObject | null;
-  headerSection?: ((resource: KubeObject | null) => React.ReactNode) | React.ReactNode;
-  title?: string;
-  extraInfo?:
-    | ((resource: KubeObject | null) => NameValueTableRow[] | null)
-    | NameValueTableRow[]
-    | null;
-  actions?:
-    | ((resource: KubeObject | null) => React.ReactNode[] | null)
-    | React.ReactNode[]
-    | null
-    | HeaderAction[];
-  headerStyle?: HeaderStyleProps['headerStyle'];
-  noDefaultActions?: boolean;
-  /** The route or location to go to. If it's an empty string, then the "browser back" function is used. If null, no back button will be shown. */
-  backLink?: string | ReturnType<typeof useLocation> | null;
-  error?: string | Error | null;
-}
-
-export function MainInfoSection(props: MainInfoSectionProps) {
-  const {
-    resource,
-    headerSection,
-    title,
-    extraInfo = [],
-    actions = [],
-    headerStyle = 'main',
-    noDefaultActions = false,
-    backLink,
-    error = null,
-  } = props;
-  const headerActions = useTypedSelector(state => state.ui.views.details.headerActions);
-  const headerActionsProcessors = useTypedSelector(
-    state => state.ui.views.details.headerActionsProcessors
-  );
-  const { t } = useTranslation('frequent');
-  const header = typeof headerSection === 'function' ? headerSection(resource) : headerSection;
-
-  function setupAction(headerAction: HeaderAction) {
-    let Action = headerAction.action;
-    if (!Action && !noDefaultActions) {
-      switch (headerAction.id) {
-        case DefaultHeaderAction.RESTART:
-          Action = RestartButton;
-          break;
-        case DefaultHeaderAction.SCALE:
-          Action = ScaleButton;
-          break;
-        case DefaultHeaderAction.EDIT:
-          Action = EditButton;
-          break;
-        case DefaultHeaderAction.DELETE:
-          Action = DeleteButton;
-          break;
-        default:
-          break;
-      }
-    }
-
-    if (!Action) {
-      return null;
-    }
-
-    if (isValidElement(Action)) {
-      return <ErrorBoundary>{Action}</ErrorBoundary>;
-    } else if (Action === null) {
-      return null;
-    } else if (typeof Action === 'function') {
-      return (
-        <ErrorBoundary>
-          <Action item={resource} />
-        </ErrorBoundary>
-      );
-    }
-  }
-
-  const defaultActions = [
-    {
-      id: DefaultHeaderAction.RESTART,
-    },
-    {
-      id: DefaultHeaderAction.SCALE,
-    },
-    {
-      id: DefaultHeaderAction.EDIT,
-    },
-    {
-      id: DefaultHeaderAction.DELETE,
-    },
-  ];
-
-  let hAccs: HeaderAction[] = [];
-  const accs = typeof actions === 'function' ? actions(resource) || [] : actions;
-  if (accs !== null) {
-    hAccs = [...accs].map((action, i) => {
-      if ((action as HeaderAction).id !== undefined) {
-        return action as HeaderAction;
-      } else {
-        return { id: `gen-${i}`, action: () => action };
-      }
-    });
-  }
-
-  let actionsProcessed = [...headerActions, ...hAccs, ...defaultActions];
-  if (headerActionsProcessors.length > 0) {
-    for (const headerProcessor of headerActionsProcessors) {
-      actionsProcessed = headerProcessor.processor(resource, actionsProcessed);
-    }
-  }
-
-  const allActions = React.Children.toArray(
-    (function propsActions() {
-      const pluginAddedActions = actionsProcessed.map(setupAction);
-      return React.Children.toArray(pluginAddedActions);
-    })()
-  );
-
-  function getBackLink() {
-    if (!!backLink || backLink === '') {
-      return backLink;
-    }
-
-    if (!!resource) {
-      return createRouteURL(resource.listRoute);
-    }
-  }
-
-  return (
-    <SectionBox
-      aria-busy={resource === null}
-      aria-live="polite"
-      title={
-        <SectionHeader
-          title={title || (resource ? resource.kind : '')}
-          headerStyle={headerStyle}
-          actions={allActions}
-        />
-      }
-      backLink={getBackLink()}
-    >
-      {resource === null ? (
-        !!error ? (
-          <Empty color="error">{error.toString()}</Empty>
-        ) : (
-          <Loader title={t('frequent|Loading resource data')} />
-        )
-      ) : (
-        <React.Fragment>
-          {header}
-          <MetadataDisplay resource={resource} extraRows={extraInfo} />
-        </React.Fragment>
-      )}
-    </SectionBox>
   );
 }
